fix: rename providers package to policy
