+++ policy/doc.go
@@ -1,0 +1,7 @@
+/*
+ Copyright 2024 The OSCAL Compass Authors
+ SPDX-License-Identifier: Apache-2.0
+*/
+
+// Package policy contains types that define the policy plugin protocol.
+package policy
