+++ policy/provider.go
@@ -3,18 +3,18 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-package providers
+package policy
 
 import "github.com/oscal-compass/oscal-sdk-go/extensions"
 
 /*
-PolicyProvider defines methods for a policy validation engine.
+Provider defines methods for a policy validation engine.
 Defined uses cases include the following:
 1. A scanning plugin may contact a remote API for scanning
 2. A scanning plugin may exec out to another tool for scanning in a new process
 3. A scanning plugin may be a self-contained scanning tool
 */
-type PolicyProvider interface {
+type Provider interface {
 	Generate(rules []extensions.RuleSet) error
 	GetResults(rules []extensions.RuleSet) (PVPResult, error)
 }
