+++ policy/types.go
@@ -3,7 +3,7 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-package providers
+package policy
 
 import (
 	"time"
