--- providers/doc.go
@@ -1,7 +1,0 @@
-/*
- Copyright 2024 The OSCAL Compass Authors
- SPDX-License-Identifier: Apache-2.0
-*/
-
-// Package providers contains interface that define supported plugin protocols.
-package providers
